+++ src/site/markdown/release-notes.md
@@ -5,6 +5,7 @@
   * Add optional preamble HTML to login page, defined in lang.json (issue #415)
   * Add showAsButton to authenticator configuration (in topcat.json) to give an authenticator its own login button (issue #418)
   * Add extraLoginButtons to facility configuration (in topcat.json) (issue #419)
+  * Remove DOI and some login redirections from browser history (issue #424)
   
 ## 2.4.3 (25th Mar 2019)